refactor: move Permissions.ts to models
