+++ src/containers/HospitalRun.tsx
@@ -3,7 +3,7 @@ import { Switch, Route } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { Toaster } from '@hospitalrun/components'
 import Sidebar from '../components/Sidebar'
-import Permissions from '../util/Permissions'
+import Permissions from '../model/Permissions'
 import Dashboard from './Dashboard'
 import Patients from '../patients/list/Patients'
 import NewPatient from '../patients/new/NewPatient'
