+++ src/slices/user-slice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import Permissions from '../util/Permissions'
+import Permissions from '../model/Permissions'
 
 interface UserState {
   permissions: Permissions[]
